Formatting stuff, fix typo in NonScrollableListView
