--- app/src/main/java/com/example/musicmap/customUI/view/NonScrollabelListView.java
@@ -1,29 +1,0 @@
-package com.example.musicmap.customUI.view;
-
-import android.content.Context;
-import android.util.AttributeSet;
-import android.view.ViewGroup;
-import android.widget.ListView;
-
-public class NonScrollabelListView extends ListView {
-
-    public NonScrollabelListView(Context context) {
-        super(context);
-    }
-
-    public NonScrollabelListView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    public NonScrollabelListView(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-    }
-
-    @Override
-    public void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        int heightMeasureSpec_custom = MeasureSpec.makeMeasureSpec(Integer.MAX_VALUE >> 2, MeasureSpec.AT_MOST);
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec_custom);
-        ViewGroup.LayoutParams params = getLayoutParams();
-        params.height = getMeasuredHeight();
-    }
-}
