+++ app/src/main/java/com/example/musicmap/customUI/view/NonScrollableListView.java
@@ -1,0 +1,33 @@
+package com.example.musicmap.customUI.view;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.ViewGroup;
+import android.widget.ListView;
+
+/**
+ * A {@link ListView} that cannot be scrolled through.
+ */
+public class NonScrollableListView extends ListView {
+
+    public NonScrollableListView(Context context) {
+        super(context);
+    }
+
+    public NonScrollableListView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public NonScrollableListView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+    }
+
+    @Override
+    public void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        heightMeasureSpec = MeasureSpec.makeMeasureSpec(Integer.MAX_VALUE >> 2, MeasureSpec.AT_MOST);
+        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
+        ViewGroup.LayoutParams params = getLayoutParams();
+        params.height = getMeasuredHeight();
+    }
+
+}
