+++ app/src/main/res/layout/fragment_profile_page.xml
@@ -94,7 +94,7 @@
                         app:layout_constraintStart_toStartOf="parent"
                         app:layout_constraintTop_toTopOf="parent" />
 
-                    <com.example.musicmap.customUI.view.NonScrollabelListView
+                    <com.example.musicmap.customUI.view.NonScrollableListView
                         android:id="@+id/mm_list"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
